ratings and friendship
The rating system is now a field of the book review model so users must fill out a review as well as a rating for books. Friendship model has been added and will be used to determine who users can interact with based on friendship status. Views have been modified so they all inherit from the the APIview for consistency and readability.