avatar, friend request, serializer changes
User model now contains an avatar field with a default avatar and can be changed and deleted by users. friend request view and serializer have been created though they are not configured on the frontend and have not been tested. Deleted requestUserSerializer as it performed the same function as the NookUserSerializer. Review serializer now sends over the user data from the NookUserSerializer to give users access to the data of the users who wrote the reviews.