review view and friendship serializer
creating and deleting reviews is now done in the same view using post and delete methods and the is_owner field is added back to the review serializer since its easier to do it on the backend than the frontend. Fixed issue in friendship view where the data was being passed through the serializer twice.