Thread create
Threads can now be created from the frontend. A thread serializer has been set up with functions for computing the name of the thread and the avatar that the thread should use. A signals.py file has been set up to automatically update the name of the thread upon creation or patching since the save method cannot use many to many values.